fix(http): replace shorthand for named route Plus minor refactor for shorthand replacement methods
